Add endpoints for device management
- Add missing log messages and validations to UserEndpoint

+++ server-webapp/src/main/java/ar/edu/itba/iot/carne_iot/server/web/controller/rest_endpoints/UserEndpoint.java
@@ -1,11 +1,14 @@
 package ar.edu.itba.iot.carne_iot.server.web.controller.rest_endpoints;
 
-import ar.edu.itba.iot.carne_iot.server.error_handling.helpers.ValidationExceptionThrower;
+import ar.edu.itba.iot.carne_iot.server.models.Device;
 import ar.edu.itba.iot.carne_iot.server.models.Role;
 import ar.edu.itba.iot.carne_iot.server.models.User;
+import ar.edu.itba.iot.carne_iot.server.services.DeviceService;
 import ar.edu.itba.iot.carne_iot.server.services.UserService;
 import ar.edu.itba.iot.carne_iot.server.web.controller.dtos.entities.StringValueDto;
+import ar.edu.itba.iot.carne_iot.server.web.controller.dtos.entities.UserDeviceDto;
 import ar.edu.itba.iot.carne_iot.server.web.controller.dtos.entities.UserDto;
+import ar.edu.itba.iot.carne_iot.server.web.support.annotations.Base64url;
 import ar.edu.itba.iot.carne_iot.server.web.support.annotations.Java8Time;
 import ar.edu.itba.iot.carne_iot.server.web.support.annotations.JerseyController;
 import ar.edu.itba.iot.carne_iot.server.web.support.annotations.PaginationParam;
@@ -19,10 +22,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 import javax.ws.rs.*;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.UriInfo;
+import javax.ws.rs.core.*;
 import java.net.URI;
 import java.time.LocalDate;
 import java.util.*;
@@ -35,7 +35,7 @@ import java.util.stream.Collectors;
 @Path(UserEndpoint.USERS_ENDPOINT)
 @Produces(MediaType.APPLICATION_JSON)
 @JerseyController
-public class UserEndpoint implements ValidationExceptionThrower {
+public class UserEndpoint {
 
     /**
      * Endpoint for {@link User} management.
@@ -55,10 +55,15 @@ public class UserEndpoint implements ValidationExceptionThrower {
      */
     private final UserService userService;
 
+    /**
+     * The {@link DeviceService}.
+     */
+    private final DeviceService deviceService;
 
     @Autowired
-    public UserEndpoint(UserService userService) {
+    public UserEndpoint(UserService userService, DeviceService deviceService) {
         this.userService = userService;
+        this.deviceService = deviceService;
     }
 
 
@@ -86,6 +91,7 @@ public class UserEndpoint implements ValidationExceptionThrower {
 
         final Page<User> users = userService
                 .findMatching(fullName, minBirthDate, maxBirthDate, username, email, pageable);
+
         return Response.ok(users.getContent().stream()
                 .map(user -> new UserDto(user, getLocationUri(user, uriInfo)))
                 .collect(Collectors.toList()))
@@ -95,6 +101,10 @@ public class UserEndpoint implements ValidationExceptionThrower {
     @GET
     @Path("{id : \\d+}")
     public Response getUserById(@PathParam("id") final long id) {
+        if (id <= 0) {
+            throw new IllegalParamValueException(Collections.singletonList("id"));
+        }
+
         LOGGER.debug("Getting user by id {}", id);
 
         return getUserBySomePropertyResponse(userService.getById(id), uriInfo);
@@ -103,6 +113,10 @@ public class UserEndpoint implements ValidationExceptionThrower {
     @GET
     @Path("username/{username : .+}")
     public Response getUserByUsername(@PathParam("username") final String username) {
+        if (username == null) {
+            throw new IllegalParamValueException(Collections.singletonList("username"));
+        }
+
         LOGGER.debug("Getting user by username {}", username);
 
         return getUserBySomePropertyResponse(userService.getByUsername(username), uriInfo);
@@ -111,6 +125,10 @@ public class UserEndpoint implements ValidationExceptionThrower {
     @GET
     @Path("email/{email : .+}")
     public Response getUserByEmail(@PathParam("email") final String email) {
+        if (email == null) {
+            throw new IllegalParamValueException(Collections.singletonList("email"));
+        }
+
         LOGGER.debug("Getting user by email {}", email);
 
         return getUserBySomePropertyResponse(userService.getByEmail(email), uriInfo);
@@ -120,8 +138,12 @@ public class UserEndpoint implements ValidationExceptionThrower {
     @Consumes(MediaType.APPLICATION_JSON)
     public Response register(final UserDto userDto) {
         return Optional.ofNullable(userDto)
-                .map(dto -> userService.register(userDto.getFullName(), userDto.getBirthDate(),
-                        userDto.getUsername(), userDto.getEmail(), userDto.getPassword()))
+                .map(dto -> {
+                    LOGGER.debug("Creating user with username {}, and email {}",
+                            userDto.getUsername(), userDto.getEmail());
+                    return userService.register(userDto.getFullName(), userDto.getBirthDate(),
+                            userDto.getUsername(), userDto.getEmail(), userDto.getPassword());
+                })
                 .map(user -> Response
                         .created(uriInfo.getAbsolutePathBuilder().path(String.valueOf(user.getId())).build())
                         .build())
@@ -135,8 +157,10 @@ public class UserEndpoint implements ValidationExceptionThrower {
         if (id <= 0) {
             throw new IllegalParamValueException(Collections.singletonList("id"));
         }
+
         return Optional.ofNullable(userDto)
                 .map(dto -> {
+                    LOGGER.debug("Updating user with id {}", id);
                     userService.update(id, userDto.getFullName(), userDto.getBirthDate());
                     return Response.noContent().build();
                 })
@@ -151,9 +175,12 @@ public class UserEndpoint implements ValidationExceptionThrower {
         if (id <= 0) {
             throw new IllegalParamValueException(Collections.singletonList("id"));
         }
+
         return Optional.ofNullable(newUsernameDto)
-                .map(dto -> {
-                    userService.changeUsername(id, dto.getValue());
+                .map(StringValueDto::getValue)
+                .map(username -> {
+                    LOGGER.debug("Changing username to {} to user with id {}", username, id);
+                    userService.changeUsername(id, username);
                     return Response.noContent().build();
                 })
                 .orElseThrow(MissingJsonException::new);
@@ -166,9 +193,12 @@ public class UserEndpoint implements ValidationExceptionThrower {
         if (id <= 0) {
             throw new IllegalParamValueException(Collections.singletonList("id"));
         }
+
         return Optional.ofNullable(newEmailDto)
-                .map(dto -> {
-                    userService.changeEmail(id, dto.getValue());
+                .map(StringValueDto::getValue)
+                .map(email -> {
+                    LOGGER.debug("Changing username to {} to user with id {}", email, id);
+                    userService.changeEmail(id, email);
                     return Response.noContent().build();
                 })
                 .orElseThrow(MissingJsonException::new);
@@ -181,8 +211,10 @@ public class UserEndpoint implements ValidationExceptionThrower {
         if (id <= 0) {
             throw new IllegalParamValueException(Collections.singletonList("id"));
         }
+
         return Optional.ofNullable(passwordChangeDto)
                 .map(dto -> {
+                    LOGGER.debug("Changing password to user with id {} ", id);
                     userService.changePassword(id, dto.getCurrentPassword(), dto.getNewPassword());
                     return Response.noContent().build();
                 })
@@ -195,7 +227,11 @@ public class UserEndpoint implements ValidationExceptionThrower {
         if (id <= 0) {
             throw new IllegalParamValueException(Collections.singletonList("id"));
         }
+
+        LOGGER.debug("Getting authorities for user with id {} ", id);
+
         final Set<Role> roles = userService.getRoles(id);
+
         return Response.ok(roles).build();
     }
 
@@ -203,6 +239,9 @@ public class UserEndpoint implements ValidationExceptionThrower {
     @Path("{id : \\d+}/authorities/{role: .+}")
     public Response addAuthority(@PathParam("id") final long id, @PathParam("role") final Role role) {
         validateRoleParams(id, role);
+
+        LOGGER.debug("Adding role {} to user with id {} ", role, id);
+
         userService.addRole(id, role);
 
         return Response.noContent().build();
@@ -212,6 +251,9 @@ public class UserEndpoint implements ValidationExceptionThrower {
     @Path("{id : \\d+}/authorities/{role: .+}")
     public Response removeAuthority(@PathParam("id") final long id, @PathParam("role") final Role role) {
         validateRoleParams(id, role);
+
+        LOGGER.debug("Removing role {} to user with id {} ", role, id);
+
         userService.removeRole(id, role);
 
         return Response.noContent().build();
@@ -224,7 +266,11 @@ public class UserEndpoint implements ValidationExceptionThrower {
         if (id <= 0) {
             throw new IllegalParamValueException(Collections.singletonList("id"));
         }
+
+        LOGGER.debug("Removing user with id {}", id);
+
         userService.deleteById(id);
+
         return Response.noContent().build();
     }
 
@@ -232,7 +278,14 @@ public class UserEndpoint implements ValidationExceptionThrower {
     @Path("username/{username : .+}")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response deleteByUsername(@PathParam("username") final String username) {
+        if (username == null) {
+            throw new IllegalParamValueException(Collections.singletonList("username"));
+        }
+
+        LOGGER.debug("Removing user with username {}", username);
+
         userService.deleteByUsername(username);
+
         return Response.noContent().build();
     }
 
@@ -240,7 +293,102 @@ public class UserEndpoint implements ValidationExceptionThrower {
     @Path("email/{email : .+}")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response deleteByEmail(@PathParam("email") final String email) {
+        if (email == null) {
+            throw new IllegalParamValueException(Collections.singletonList("username"));
+        }
+
+        LOGGER.debug("Removing user with username {}", email);
+
         userService.deleteByEmail(email);
+
+        return Response.noContent().build();
+    }
+
+    // ======================================
+    // User devices
+    // ======================================
+
+    @GET
+    @Path("/{id : \\d+}" + DevicesEndpoint.DEVICES_ENDPOINT)
+    public Response listUserDevices(@SuppressWarnings("RSReferenceInspection") @PathParam("id") final long id,
+                                    @PaginationParam final Pageable pageable) {
+        if (id <= 0) {
+            throw new IllegalParamValueException(Collections.singletonList("id"));
+        }
+
+        LOGGER.debug("Listing devices belonging to user with id {}", id);
+
+        final Page<DeviceService.DeviceWithNicknameWrapper> devices = deviceService.listUserDevices(id, pageable);
+
+        return Response.ok(devices.getContent().stream()
+                .map(wrapper ->
+                        new UserDeviceDto(wrapper, getUserDeviceLocationUri(id, wrapper.getDevice(), uriInfo)))
+                .collect(Collectors.toList()))
+                .build();
+    }
+
+    @GET
+    @Path("/{id : \\d+}" + DevicesEndpoint.DEVICES_ENDPOINT + "/{deviceId : .+}")
+    public Response getDevice(@SuppressWarnings("RSReferenceInspection") @PathParam("id") final long id,
+                              @SuppressWarnings("RSReferenceInspection")
+                              @PathParam("deviceId") @Base64url final Long deviceId) {
+        validateUserDeviceParams(id, deviceId);
+
+        LOGGER.debug("Getting device with id {} belonging to user with id {}", deviceId, id);
+
+        final Optional<DeviceService.DeviceWithNicknameWrapper> wrapperOptional =
+                deviceService.getRegisteredDevice(id, deviceId);
+
+        return wrapperOptional.map(wrapper ->
+                Response.ok(new UserDeviceDto(wrapper, getUserDeviceLocationUri(id, wrapper.getDevice(), uriInfo))))
+                .orElse(Response.status(Response.Status.NOT_FOUND).entity(""))
+                .build();
+    }
+
+    @PUT
+    @Path("/{id : \\d+}" + DevicesEndpoint.DEVICES_ENDPOINT + "/{deviceId : .+}")
+    public Response registerDevice(@SuppressWarnings("RSReferenceInspection") @PathParam("id") final long id,
+                                   @SuppressWarnings("RSReferenceInspection")
+                                   @PathParam("deviceId") @Base64url final Long deviceId) {
+        validateUserDeviceParams(id, deviceId);
+
+        LOGGER.debug("Registering device with id {} belonging to user with id {}", deviceId, id);
+
+        deviceService.registerDevice(id, deviceId);
+
+        return Response.noContent().build();
+    }
+
+    @PUT
+    @Path("/{id : \\d+}" + DevicesEndpoint.DEVICES_ENDPOINT + "/{deviceId : .+}/nickname")
+    public Response changeNickname(@SuppressWarnings("RSReferenceInspection") @PathParam("id") final long id,
+                                   @SuppressWarnings("RSReferenceInspection")
+                                   @PathParam("deviceId") @Base64url final Long deviceId,
+                                   final StringValueDto stringValueDto) {
+        validateUserDeviceParams(id, deviceId);
+        if (stringValueDto == null) {
+            throw new MissingJsonException();
+        }
+
+        LOGGER.debug("Changing nickname to {} to device with id {} belonging to user with id {}",
+                stringValueDto.getValue(), deviceId, id);
+
+        deviceService.setNickname(id, deviceId, stringValueDto.getValue());
+
+        return Response.noContent().build();
+    }
+
+    @DELETE
+    @Path("/{id : \\d+}" + DevicesEndpoint.DEVICES_ENDPOINT + "/{deviceId : .+}/nickname")
+    public Response deleteNickname(@SuppressWarnings("RSReferenceInspection") @PathParam("id") final long id,
+                                   @SuppressWarnings("RSReferenceInspection")
+                                   @PathParam("deviceId") @Base64url final Long deviceId) {
+        validateUserDeviceParams(id, deviceId);
+
+        LOGGER.debug("Removing nickname to device with id {} belonging to user with id {}", deviceId, id);
+
+        deviceService.deleteNickname(id, deviceId);
+
         return Response.noContent().build();
     }
 
@@ -265,6 +413,20 @@ public class UserEndpoint implements ValidationExceptionThrower {
     }
 
     /**
+     * Returns a {@link UriBuilder} for a given {@link User} location.
+     *
+     * @param userId  The id of {@link User}'s whose location {@link UriBuilder} must be retrieved.
+     * @param uriInfo The {@link UriInfo} holding the context.
+     * @return The initialized {@link UriBuilder}.
+     */
+    /* package */
+    static UriBuilder baseUserUriBuilder(long userId, UriInfo uriInfo) {
+        return uriInfo.getBaseUriBuilder().clone()
+                .path(USERS_ENDPOINT)
+                .path(Long.toString(userId));
+    }
+
+    /**
      * Returns the location {@link URI} of the given {@link User}
      * according to the context hold by the given {@link UriInfo}
      *
@@ -273,10 +435,43 @@ public class UserEndpoint implements ValidationExceptionThrower {
      * @return The location {@link URI} of the given {@link User}
      */
     private static URI getLocationUri(User user, UriInfo uriInfo) {
-        return uriInfo.getBaseUriBuilder().clone()
-                .path(USERS_ENDPOINT)
-                .path(Long.toString(user.getId()))
+        return baseUserUriBuilder(user.getId(), uriInfo)
                 .build();
+    }
+
+    /**
+     * Returns the location {@link URI} of the given {@link Device} of the given {@link User},
+     * according to the context hold by the given {@link UriInfo}
+     *
+     * @param userId  The id of {@link User}'s whose location {@link URI} must be retrieved.
+     * @param device  The {@link Device}'s whose location {@link URI} must be retrieved.
+     * @param uriInfo The {@link UriInfo} holding the context.
+     * @return The location {@link URI} of the given {@link User}
+     */
+    private static URI getUserDeviceLocationUri(long userId, Device device, UriInfo uriInfo) {
+        return baseUserUriBuilder(userId, uriInfo)
+                .path(DevicesEndpoint.DEVICES_ENDPOINT)
+                .path(DevicesEndpoint.toBase64UrlId(device))
+                .build();
+    }
+
+    /**
+     * Performs validation over the given params.
+     *
+     * @param userId   The {@link User} id param to be validated.
+     * @param deviceId The {@link Device} id param to be validated.
+     */
+    private void validateUserDeviceParams(long userId, Long deviceId) {
+        final List<String> wrongParams = new LinkedList<>();
+        if (userId <= 0) {
+            wrongParams.add("id");
+        }
+        if (deviceId == null) {
+            wrongParams.add("deviceId");
+        }
+        if (!wrongParams.isEmpty()) {
+            throw new IllegalParamValueException(wrongParams);
+        }
     }
 
     /**
